+++ src/PostgREST/Parsers.hs
@@ -134,12 +134,12 @@ pRelationSelect :: Parser SelectItem
 pRelationSelect = lexeme $ try ( do
     alias <- optionMaybe ( try(pFieldName <* aliasSeparator) )
     fld <- pField
-    relationDetail <- optionMaybe (
-        try ( char '!' *> pFieldName ) <|>
-        try ( char '.' *> pFieldName ) -- TODO deprecated, remove in next major version
+    hint <- optionMaybe (
+        try ( char '!' *> pFieldName) <|>
+        -- deprecated, remove in next major version
+        try ( char '.' *> pFieldName)
       )
-
-    return (fld, Nothing, alias, relationDetail)
+    return (fld, Nothing, alias, hint)
   )
 
 pFieldSelect :: Parser SelectItem