+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - #1383, Add support for HEAD request - @steve-chavez
 - #1378, Add support for `Prefer: count=planned` and `Prefer: count=estimated` on GET /table - @steve-chavez
 - #1327, Add support for optional query parameter `on_conflict` to upsert with specified keys for POST - @ykst
+- #1430, Allow specifying the constraint name(`/source?select=constraint`) to disambiguate an embedding - @steve-chavez
 
 ### Fixed
 
@@ -25,7 +26,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - #1385, bulk RPC call now should be done by specifying a `Prefer: params=multiple-objects` header - @steve-chavez
 - #1401, resource embedding now outputs an error when multiple relationships between two tables are found - @steve-chavez
 - #1423, default Unix Socket file mode from 755 to 660 - @dwagin
-- Remove embedding with duck typed column names `GET /projects?select=client(*)`- @steve-chavez
+- #1430, Remove embedding with duck typed column names `GET /projects?select=client(*)`- @steve-chavez
+  + You can rename the foreign key to `client` to make this request work in the new version: `alter table projects rename constraint projects_client_id_fkey to client`
 
 ## [6.0.2] - 2019-08-22